fix: rename bounded name
