--- src/main/java/com/IoTeam/ThirstySeedAPI/datamanagement/plot/domain/model/commands/CreatePlotCommand.java
@@ -1,13 +1,0 @@
-package com.IoTeam.ThirstySeedAPI.datamanagement.plot.domain.model.commands;
-
-import java.util.List;
-
-public record CreatePlotCommand(
-        String name,
-        String location,
-        Integer extension,
-        Integer size,
-        String imageUrl
-
-) {
-}
