--- src/main/java/com/IoTeam/ThirstySeedAPI/datamanagement/plot/domain/model/commands/NotSupplyPlotCommand.java
@@ -1,4 +1,0 @@
-package com.IoTeam.ThirstySeedAPI.datamanagement.plot.domain.model.commands;
-
-public record NotSupplyPlotCommand(Long plotId) {
-}
