--- src/main/java/com/IoTeam/ThirstySeedAPI/datamanagement/plot/domain/model/queries/GetAllPlotsQuery.java
@@ -1,4 +1,0 @@
-package com.IoTeam.ThirstySeedAPI.datamanagement.plot.domain.model.queries;
-
-public record GetAllPlotsQuery() {
-}
