--- src/main/java/com/IoTeam/ThirstySeedAPI/datamanagement/plot/domain/model/queries/GetPlotByIdQuery.java
@@ -1,4 +1,0 @@
-package com.IoTeam.ThirstySeedAPI.datamanagement.plot.domain.model.queries;
-
-public record GetPlotByIdQuery(Long idPlot) {
-}
