--- src/main/java/com/IoTeam/ThirstySeedAPI/datamanagement/plot/domain/model/valueobjects/Location.java
@@ -1,19 +1,0 @@
-package com.IoTeam.ThirstySeedAPI.datamanagement.plot.domain.model.valueobjects;
-
-import jakarta.persistence.Embeddable;
-
-@Embeddable
-public record Location(String location) {
-    public Location(){
-        this(null);
-    }
-
-    public Location {
-        if (location == null || location.isBlank()) {
-            throw new IllegalArgumentException("Location cannot be null or blank");
-        }
-
-    }
-
-
-}
