--- src/main/java/com/IoTeam/ThirstySeedAPI/datamanagement/plot/domain/model/valueobjects/Name.java
@@ -1,19 +1,0 @@
-package com.IoTeam.ThirstySeedAPI.datamanagement.plot.domain.model.valueobjects;
-
-import jakarta.persistence.Embeddable;
-
-@Embeddable
-public record Name(String name) {
-    public Name(){
-        this(null);
-    }
-
-    public Name {
-        if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException("Name cannot be null or blank");
-        }
-
-    }
-
-
-}
