--- src/main/java/com/IoTeam/ThirstySeedAPI/datamanagement/plot/interfaces/resources/CreatePlotResource.java
@@ -1,11 +1,0 @@
-package com.IoTeam.ThirstySeedAPI.datamanagement.plot.interfaces.resources;
-
-public record CreatePlotResource(
-
-       String name,
-       String location,
-       Integer extension,
-       Integer size,
-       String imageUrl
-) {
-}
