+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/exceptions/PlotNotExceptions.java
@@ -1,0 +1,7 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.exceptions;
+
+public class PlotNotExceptions extends RuntimeException{
+     public PlotNotExceptions(Long plotId){
+         super("Plot not found with id: " + plotId);
+    }
+}
