+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/model/commands/CreatePlotCommand.java
@@ -1,0 +1,11 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.model.commands;
+
+public record CreatePlotCommand(
+        String name,
+        String location,
+        Integer extension,
+        Integer size,
+        String imageUrl
+
+) {
+}
