+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/model/commands/NotSupplyPlotCommand.java
@@ -1,0 +1,4 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.model.commands;
+
+public record NotSupplyPlotCommand(Long plotId) {
+}
