+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/model/commands/SupplyPlotCommand.java
@@ -1,0 +1,4 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.model.commands;
+
+public record SupplyPlotCommand(Long plotId) {
+}
