+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/model/queries/GetAllPlotsQuery.java
@@ -1,0 +1,4 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.model.queries;
+
+public record GetAllPlotsQuery() {
+}
