+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/model/valueobjects/Extension.java
@@ -1,0 +1,18 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.model.valueobjects;
+
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+public record Extension(Integer extension) {
+
+    public Extension() {
+        this(0);
+    }
+
+    public Extension {
+        if (extension == null || extension <= 0) {
+            throw new IllegalArgumentException("Extension value must be more than zero");
+        }
+    }
+}
