+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/model/valueobjects/Location.java
@@ -1,0 +1,19 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.model.valueobjects;
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+public record Location(String location) {
+    public Location(){
+        this(null);
+    }
+
+    public Location {
+        if (location == null || location.isBlank()) {
+            throw new IllegalArgumentException("Location cannot be null or blank");
+        }
+
+    }
+
+
+}
