+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/model/valueobjects/Name.java
@@ -1,0 +1,19 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.model.valueobjects;
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+public record Name(String name) {
+    public Name(){
+        this(null);
+    }
+
+    public Name {
+        if (name == null || name.isBlank()) {
+            throw new IllegalArgumentException("Name cannot be null or blank");
+        }
+
+    }
+
+
+}
