+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/domain/model/valueobjects/PlotStatus.java
@@ -1,0 +1,7 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.domain.model.valueobjects;
+
+public enum PlotStatus {
+    REQUESTED,
+    SUPPLIED,
+    NOT_SUPPLIED
+}
