+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/infrastructure/persistence/jpa/repositories/PlotRepository.java
@@ -1,0 +1,10 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.infrastructure.persistence.jpa.repositories;
+
+import com.IoTeam.ThirstySeedAPI.irrigation.domain.model.aggregates.Plot;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PlotRepository extends JpaRepository<Plot,Long> {
+
+
+}
