+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/interfaces/resources/CreatePlotResource.java
@@ -1,0 +1,11 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.interfaces.resources;
+
+public record CreatePlotResource(
+
+       String name,
+       String location,
+       Integer extension,
+       Integer size,
+       String imageUrl
+) {
+}
