+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/interfaces/resources/PlotResource.java
@@ -1,0 +1,18 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.interfaces.resources;
+
+import java.util.Date;
+
+public record PlotResource(
+        Long id,
+        String name,
+        String location,
+        Integer extension,
+        Integer size,
+        String status,
+        String imageUrl,
+        Date createdAt,
+        Date updatedAt
+) {
+
+
+}
