+++ src/main/java/com/IoTeam/ThirstySeedAPI/irrigation/interfaces/transform/CreatePlotCommandFromResourceAssembler.java
@@ -1,0 +1,16 @@
+package com.IoTeam.ThirstySeedAPI.irrigation.interfaces.transform;
+
+import com.IoTeam.ThirstySeedAPI.irrigation.domain.model.commands.CreatePlotCommand;
+import com.IoTeam.ThirstySeedAPI.irrigation.interfaces.resources.CreatePlotResource;
+
+public class CreatePlotCommandFromResourceAssembler {
+    public static CreatePlotCommand toCommandFromResource(CreatePlotResource resource){
+        return new CreatePlotCommand(
+                resource.name(),
+                resource.location(),
+                resource.extension(),
+                resource.size(),
+                resource.imageUrl()
+        );
+    }
+}
